Allow element rendered by Provider to be configured
The default `<div>` wrapper isn't valid anywhere in the DOM. By allowing
the wrapper element to be configured it is possible for the user to
change it to something else when needed.

An example use case for this are component libraries with bundled
translations. If a component is intended to be used as phrasing content
then it can't introduce a `<div>`. Moving the `Provider` higher up
the tree isn't possible here, as this falls out of the scope of the
library.